Worked on the sidedar. Moved it to the left due to issues with the web page. also added a back button for the drill page
